2005-04-12  Aaron Luchko  <aluchko@redhat.com>
	PR libgcj/20958
	* gnu/gcj/runtime/PersistentByteMap.java (close): Throw IOException.
	* gnu/gcj/tools/gcj_dbtool/Main.java (addJar): Make 'length' an int.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@98053 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libjava/gnu/gcj/tools/gcj_dbtool/Main.java
@@ -375,7 +375,7 @@ public class Main
 	if (classfile.getName().endsWith(".class"))
 	  {
 	    InputStream str = jar.getInputStream(classfile);
-	    long length = classfile.getSize();
+	    int length = (int) classfile.getSize();
 	    if (length == -1)
 	      throw new EOFException();
 
@@ -383,7 +383,7 @@ public class Main
 	    int pos = 0;
 	    while (length - pos > 0)
 	      {
-		int len = str.read(data, pos, (int)(length - pos));
+		int len = str.read(data, pos, length - pos);
 		if (len == -1)
 		  throw new EOFException("Not enough data reading from: "
 					 + classfile.getName());
